requirements.txt
# LangChain + LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# PDF parsing
pdfplumber>=0.11.0       # structured PDF extraction (tables, columns)
pypdf>=4.0.0             # fallback for simple PDFs
pytesseract>=0.3.10      # OCR for scanned PDFs
Pillow>=10.0.0           # image handling for OCR

# Data validation
pydantic>=2.0.0          # structured LLM outputs

# Environment
python-dotenv>=1.0.0     # load .env file

# CLI
rich>=13.0.0             # pretty terminal output (tables, progress, markdown)
typer>=0.12.0            # clean CLI interface

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0